+++ sources/src/main/java/com/google/solutions/jitaccess/core/auth/GroupEmail.java
@@ -73,7 +73,7 @@ public class GroupEmail implements Comparable<GroupEmail>, PrincipalIdentifier {
   }
 
   // -------------------------------------------------------------------------
-  // Principal.
+  // PrincipalIdentifier.
   // -------------------------------------------------------------------------
 
   @Override
